Saves parent info in account
-----------------------TO-DO-----------------------
My powered milk for babies -> pin them and show it in the kid's milk
fragment_profile must be redesigned.
Fragment_rules must be completed.
watch out for consistencies in buttons and headers.
fragment_parent_info is redesigned but it is very similar to fragment_parent_show.
Milk result is not correct.